backend/requirements.txt
fastapi>=0.110.0
sqlmodel>=0.0.16
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
uvicorn[standard]>=0.27.1
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
email-validator>=2.1.1
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
# AI/Agent Dependencies
openai-agents>=0.1.0
mcp>=0.1.0
pydantic-ai>=0.0.1